backend/ai_server/requirements.txt
# Dependencies for Farmora AI Server pipeline

# Audio transcription
whisper
torch

# API and web interactions
requests
bs4
selenium
webdriver-manager
python-dotenv

# LLM/API providers
groq
huggingface-hub

# Data processing
numpy
pathlib
json

# For development
pytest